BOOKINGSG-3753 [VG] Toast component changes
- Rename component
- Fix props
- Fix color specification
