+++ src/spec/color-spec/base-color-set.ts
@@ -57,11 +57,11 @@ export const BaseColorSet: ColorSet = {
             Border: "#DD6666",
             Background: "#FDF7F7", // Validation Background
         },
-        Dark: {
-            Text: "#FFFFFF", // Validation Text
-            Icon: "#FFFFFF",
-            Border: "#424242",
-            Background: "#424242", // Validation Background
+        Blue: {
+            Text: "#064D70", // Validation Text
+            Icon: " #066391",
+            Border: "#84B7DB",
+            Background: "#E8F4FA", // Validation Background
         },
     } as ValidationTypes,
     Shadow: {
