+++ src/toast/toast.tsx
@@ -1,6 +1,5 @@
 import React, { useState } from "react";
 import { NotificationToastProps } from "./types";
-import { CrossIcon } from "@lifesg/react-icons/cross";
 import { ICircleFillIcon } from "@lifesg/react-icons/i-circle-fill";
 import { TickCircleFillIcon } from "@lifesg/react-icons/tick-circle-fill";
 import { ExclamationCircleFillIcon } from "@lifesg/react-icons/exclamation-circle-fill";
@@ -9,10 +8,12 @@ import {
     CloseIcon,
     Container,
     Description,
+    IconContainer,
     StyledIconButton,
+    TextContainer,
     Title,
     Wrapper,
-} from "./toast.styles.new";
+} from "./toast.styles";
 
 export const Notification = ({
     theme = "light",
@@ -46,7 +47,20 @@ export const Notification = ({
     return (
         <Wrapper $type={type} $theme={theme}>
             <Container>
-                {showIcon()}
+                <IconContainer>{showIcon()}</IconContainer>
+                <TextContainer>
+                    {title && (
+                        <Title $type={type} $theme={theme}>
+                            {title}
+                        </Title>
+                    )}
+
+                    {children && (
+                        <Description $theme={theme} $type={type}>
+                            {children}
+                        </Description>
+                    )}
+                </TextContainer>
                 <StyledIconButton
                     $type={type}
                     $theme={theme}
@@ -54,34 +68,7 @@ export const Notification = ({
                 >
                     <CloseIcon />
                 </StyledIconButton>
-                {title && (
-                    <Title $type={type} $theme={theme}>
-                        {title}
-                    </Title>
-                )}
-
-                {children && (
-                    <Description $theme={theme} $type={type}>
-                        {children}
-                    </Description>
-                )}
             </Container>
-            {/* <Container>
-                {displayIcon()}
-                <StyledIconButton $type={type} onClick={handleDismiss}>
-                    <CloseIcon />
-                </StyledIconButton>
-                {children && title && (
-                    <>
-                        <Title>{title}</Title>
-
-                        <Description $type={type}>{children}</Description>
-                    </>
-                )}
-                {!children && title && <>{title}</>}
-
-                {!title && children && <>{children}</>}
-            </Container> */}
         </Wrapper>
     );
 };
